% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binaryDist-package.R, R/binaryDist.R
\docType{package}
\name{binaryDist}
\alias{binaryDist}
\alias{binaryDist-package}
\title{binaryDist: Efficient binary/Jaccard distance}
\usage{
binaryDist(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\description{
binaryDist provides performs fast binary/Jaccard distance calculations on a
matrix. It uses raw vectors with the binary data efficiently packed 8 bits to
a byte and makes use of 64 bit operators.

Calculate binary distance on a matrix
}

