% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomData.R
\name{makeRandomData}
\alias{makeRandomData}
\title{Generate a random sparse matrix for testing purposes}
\usage{
makeRandomData(nObs, nVariables, maxBits, packed = FALSE)
}
\arguments{
\item{nObs}{}

\item{nVariables}{}

\item{maxBits}{}

\item{packed}{}
}
\description{
Generate a random sparse matrix for testing purposes
}

